fix: compilation failure caused by inconsistent syscalls on different systems
